requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.13.0
asttokens==3.0.1
attrs==26.1.0
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
colorama==0.4.6
comm==0.2.3
cryptography==46.0.6
dataclasses-json==0.6.7
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
executing==2.2.1
filetype==1.2.0
frozenlist==1.8.0
google-auth==2.49.1
google-genai==1.70.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
ipykernel==7.2.0
ipython==9.12.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonpatch==1.33
jsonpointer==3.1.1
jupyter_client==8.8.0
jupyter_core==5.9.1
langchain==1.2.14
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.25
langchain-google-genai==4.2.1
langchain-text-splitters==1.1.1
langgraph==1.1.4
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.24
marshmallow==3.26.2
matplotlib-inline==0.2.1
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.4.4
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
parso==0.8.6
platformdirs==4.9.4
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.2
pure_eval==0.2.3
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.33.1
requests-toolbelt==1.0.0
setuptools==82.0.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
stack-data==0.6.3
tenacity==9.1.4
tornado==6.5.5
traitlets==5.14.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
wcwidth==0.6.0
websockets==16.0
wheel==0.46.3
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
